Change log - v140730-RES02
//Changes/

- Changed the title to relbeits.NET
- Changed the online directories of jQuery and bootstrap to
a local folder

//Author notes/

I made a template to the change log id... why? Nothing special. Just a way to keep track of the changes.

The rule is the following: vYYMMDD-UUUN. It means:

v + YEAR + MONTH + DAY + FIRST, SECOND AND LAST LETTER OF THE USERNAME + THE NUMBER OF CHANGES BY THIS USERNAME ON THIS DAY


~relbeits_